feat: Tty: towards "enable all" and "debug" overlay
