+++ src/npc-cli/world/WorldMenu.jsx
@@ -6,6 +6,7 @@ import { geom } from '../service/geom';
 import { WorldContext } from "./world-context";
 import useStateRef from "../hooks/use-state-ref";
 import useUpdate from "../hooks/use-update";
+import { faderOverlayCss, pausedControlsCss } from "./overlay-menu-css";
 
 
 /**
@@ -84,7 +85,7 @@ export default function WorldMenu(props) {
         >
           debug
         </button>
-    </div>
+      </div>
     )}
   </>;
 }
@@ -113,60 +114,6 @@ const contextMenuCss = css`
     color: black;
     max-width: 100px;
     margin: 8px 0;
-  }
-`;
-
-const faderOverlayCss = css`
-  position: absolute;
-  z-index: 4;
-
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  
-  background: rgba(1, 1, 1, 1);
-  opacity: 1;
-  transition: opacity 1s ease-in;
-  &.clear {
-    pointer-events: none;
-    opacity: 0;
-    transition: opacity 0.5s ease-in;
-  }
-  &.faded {
-    cursor: pointer;
-    opacity: 0.5;
-    transition: opacity 0.5s ease-in;
-  }
-`;
-
-const pausedControlsCss = css`
-  position: absolute;
-  right: 0;
-  top: 64px;
-  z-index: 4;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-
-  button {
-    color: #aaa;
-    padding: 8px 12px 8px 12px;
-    background-color: #000;
-    border-top-left-radius: 8px;
-    border-bottom-left-radius: 8px;
-    border-width: 1px 0 1px 1px;
-    border-color: #444;
-    font-size: 0.8rem;
-
-    &.highlight {
-      color: #0f0;
-    }
-  }
-
-  transition: filter 1s;
-  &:hover {
-    filter: brightness(2) ;
   }
 `;
